backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.0
httpx==0.28.1
langchain-google-genai>=4.2.1
